feat: Refactor header component with separate Navbar and AuxBar components
- Extract Navbar component from Header component for better organization and reusability
- Create AuxBar component to handle auxiliary elements in the header
- Improve code readability and maintainability by splitting logic into separate components
